Fixed a number of build warnings.
- Conversion from 64-bit to 32-bit types.
- Replaced deprecated -compositeToPoint:operation:fraction:.
- Added protocol that defines -preferencesDidChange:.
- Casting delegate to AppController* in MidiKeysApplication.m.

+++ PreferencesController.m
@@ -118,7 +118,7 @@ static PreferencesController *_sharedPrefsController = nil;
     [_velocityOverlaysCheckbox setIntValue:[defaults boolForKey:SHOW_VELOCITY_OVERLAYS_PREF_KEY]];
     
 	// modifier checkboxes
-	int modifiers = [defaults integerForKey:kHotKeysModifiersPrefKey];
+	long modifiers = [defaults integerForKey:kHotKeysModifiersPrefKey];
 	BOOL controlChecked = (modifiers & controlKey) > 0;
 	BOOL shiftChecked = (modifiers & shiftKey) > 0;
 	BOOL optionChecked = (modifiers & optionKey) > 0;
